// src/enums.ts
export enum EnumAppKeys {
  isLoggedIn = "isLoggedIn",
  userData = "userData",
  userJourney = "userJourney",
}

export enum EnumUserRole {
  admin = "admin",
  user = "user",
  judge = "judge",
}

export enum EnumUsersRoutes {
  admin = "/admin",
  user = "/user",
  judge = "/judge",
}

export enum EnumUserStatus {
  active = "active",
  inactive = "inactive",
  pending = "pending",
}